move navigation to modals
