+++ web/app/containers/reducers.js
@@ -15,7 +15,7 @@ import footer from './footer/reducer'
 import header from './header/reducer'
 import home from './home/reducer'
 import login from './login/reducer'
-import navigation from './navigation/reducer'
+import navigation from '../modals/navigation/reducer'
 import productDetails from './product-details/reducer'
 import productList from './product-list/reducer'
