+++ web/app/modals/navigation/_navigation.scss
@@ -4,7 +4,7 @@
 // Social Link
 // ---
 
-.t-navigation__social-link {
+.m-navigation__social-link {
     display: block;
 
     & + & {
@@ -15,7 +15,7 @@
 // Copyright
 // ---
 
-.t-navigation__copyright {
+.m-navigation__copyright {
     color: $neutral-40;
     font-size: $small-font-size;
     line-height: $font-size;
