+++ web/app/modals/navigation/reducer.js
