+++ web/app/styles/_modals.scss
@@ -7,3 +7,4 @@
 @import '../modals/product-details-item-added/product-details-item-added';
 @import '../modals/product-list-filter/product-list-filter';
 @import '../modals/remember-me/remember-me';
+@import '../modals/navigation/navigation';
