+++ web/app/styles/_templates.scss
@@ -9,7 +9,6 @@
 @import '../containers/product-list/product-list';
 @import '../containers/header/header';
 @import '../containers/footer/footer';
-@import '../containers/navigation/navigation';
 @import '../containers/login/login';
 @import '../containers/product-details/product-details';
 @import '../containers/checkout-header/checkout-header';
